+++ R/ref-grid.R
@@ -87,6 +87,10 @@
 #'   here via a character vector or named \code{list} specifying the nesting
 #'   structure. Specifying \code{nesting} overrides any nesting structure that
 #'   is automatically detected. See Details.
+#' @param covnest Logical value. If \code{TRUE}, covariates having more than
+#'   one value in the reference grid are included when auto-detecting nesting.
+#'   Set this to \code{TRUE} only if you have covariate values that logically
+#'   depend on some other factor's levels.
 #' @param offset Numeric scalar value (if a vector, only the first element is
 #'   used). This may be used to add an offset, or override offsets based on the model.
 #'   A common usage would be to specify \code{offset = 0} for a Poisson
@@ -177,6 +181,14 @@
 #' \%in\% country, city \%in\% (state*country)"}, and \code{nesting = c("state
 #' \%in\% country)", "city \%in\% (state*country)")}.
 #' 
+#' In certain unusual cases, a covariate (rather than a factor) may be nested.
+#' Support for such situations is limited to the extent that only covariate
+#' values that exactly match a value in the dataset is permitted. I recommend
+#' supplying a reference dataset in the \code{data} argument that contains the
+#' desired covariate values for the reference grid; then the nesting will be
+#' handled correctly if you specify \code{covnest = TRUE} and 
+#' \code{cov.reduce = FALSE}.
+#' 
 #' @section Predictors with subscripts and data-set references:
 #' When the fitted model contains subscripts or explicit references to data
 #' sets, the reference grid may optionally be post-processed to simplify the
@@ -260,7 +272,7 @@
 ref_grid <- function(object, at, cov.reduce = mean, mult.names, mult.levs, 
                      options = get_emm_option("ref_grid"), data, df, type, 
                      transform = c("none", "response", "mu", "unlink", "log"), 
-                     nesting, offset, ...) 
+                     nesting, covnest = FALSE, offset, ...) 
 {
     transform = match.arg(transform)
     if (!missing(df)) {
@@ -434,32 +446,32 @@ ref_grid <- function(object, at, cov.reduce = mean, mult.names, mult.levs,
     multresp = character(0) ### ??? was list()
     ylevs = misc$ylevs
     if(!is.null(ylevs)) { # have a multivariate situation
-        if (missing(mult.levs)) {
+        if (missing(mult.levs))
             mult.levs = ylevs
-            if(!missing(mult.names)) {
-                k = seq_len(min(length(ylevs), length(mult.names)))
-                names(mult.levs)[k] = mult.names[k]
-            } 
-            if (length(ylevs) > 1)
-                ylevs = list(seq_len(prod(sapply(mult.levs, length))))
-            
-            k = prod(sapply(mult.levs, length))
-            if (k != length(ylevs[[1]])) 
-                stop("supplied 'mult.levs' is of different length ",
-                     "than that of multivariate response")
-            for (nm in names(mult.levs))
-                ref.levels[[nm]] = mult.levs[[nm]]
-            multresp = names(mult.levs)
-            MF = do.call("expand.grid", mult.levs)
-            grid = merge(grid, MF)
-        }
-            
-        # add any matrices
-        for (nm in names(matlevs))
-            grid[[nm]] = matrix(rep(matlevs[[nm]], 
-                                    each=nrow(grid)), nrow=nrow(grid))
+        if(!missing(mult.names)) {
+            k = seq_len(min(length(ylevs), length(mult.names)))
+            names(mult.levs)[k] = mult.names[k]
+        } 
+        if (length(ylevs) > 1)
+            ylevs = list(seq_len(prod(sapply(mult.levs, length))))
+        
+        k = prod(sapply(mult.levs, length))
+        if (k != length(ylevs[[1]])) 
+            stop("supplied 'mult.levs' is of different length ",
+                 "than that of multivariate response")
+        for (nm in names(mult.levs))
+            ref.levels[[nm]] = mult.levs[[nm]]
+        multresp = names(mult.levs)
+        MF = do.call("expand.grid", mult.levs)
+        grid = merge(grid, MF)
+        
     }
-
+            
+    # add any matrices
+    for (nm in names(matlevs))
+        grid[[nm]] = matrix(rep(matlevs[[nm]], 
+                                each=nrow(grid)), nrow=nrow(grid))
+    
 # Here's a complication. If a numeric predictor was coerced to a factor, we had to
 # include all its levels in the reference grid, even if altered in 'at'
 # Moreover, whatever levels are in 'at' must be a subset of the unique values
@@ -505,7 +517,10 @@ ref_grid <- function(object, at, cov.reduce = mean, mult.names, mult.levs,
     ### --- Determine weights for each grid point --- (added ver.2.11), updated ver.2.14 to include weights
     if (!hasName(data, "(weights)"))
         data[["(weights)"]] = 1
-    nms = union(union(names(xlev), names(chrlev)), coerced$factors) # only factors, no covariates or mult.resp
+    if (!covnest)
+        nms = union(union(names(xlev), names(chrlev)), coerced$factors) # only factors, no covariates or mult.resp
+    else
+        nms = setdiff(names(ref.levels)[sapply(ref.levels, length) > 1], multresp) # all names (except multiv) for which there is > 1 level
     if (length(nms) == 0)
         wgt = rep(1, nrow(grid))  # all covariates; give each weight 1
     else {
@@ -571,13 +586,22 @@ ref_grid <- function(object, at, cov.reduce = mean, mult.names, mult.levs,
         options$predict.type = type
     }
     
-    if (!missing(nesting))
-        result@model.info$nesting = .parse_nest(nesting)
-    else if (!is.null(nst <- result@model.info$nesting))
+    if (!missing(nesting)) {
+        result@model.info$nesting = lst = .parse_nest(nesting)
+        if(!is.null(lst)) {
+            nms = union(names(lst), unlist(lst))
+            if(!all(nms %in% names(result@grid)))
+                stop("Nonexistent variables specified in 'nesting'")
+            result@misc$display = .find.nonempty.nests(result, nms)
+        }
+    }
+        
+    else if (!is.null(nst <- result@model.info$nesting)) {
+        result@misc$display = .find.nonempty.nests(result)
         if (get_emm_option("msg.nesting"))
             message("NOTE: A nesting structure was detected in the ",
-                    "fitted model:\n    ", .fmt.nest(nst), 
-                "\nIf this is incorrect, re-run or update with `nesting` specified")
+                    "fitted model:\n    ", .fmt.nest(nst))
+    }
 
     if(!is.null(options)) {
         options$object = result